File Sink: implemented arbitrary sample rate selection

+++ plugins/samplesink/filesink/filesinksettings.h
@@ -21,7 +21,7 @@
 
 struct FileSinkSettings {
     quint64 m_centerFrequency;
-    qint32 m_sampleRate;
+    quint64 m_sampleRate;
     quint32 m_log2Interp;
 
     FileSinkSettings();
